+++ inst/resources/init-rprofile.R
@@ -1,3 +1,3 @@
-#### -- Packrat Autoloader (version 0.4.8-22) -- ####
+#### -- Packrat Autoloader (version 0.4.8-23) -- ####
 source("packrat/init.R")
 #### -- End Packrat Autoloader -- ####